+++ testData/intentions/TtoTrueIntention/LoopCheck.after.R
@@ -1,0 +1,3 @@
+while(require(dplyr) == FALSE<caret>){
+# do something
+}