+++ testData/intentions/TtoTrueIntention/LoopCheck.before.R
@@ -1,0 +1,3 @@
+while(require(dplyr) == F<caret>){
+# do something
+}